Add authentication to stable diffusion UI
automatic1111 has support for api-auth on /sdapi resources, but not on the web ui.
Configure authelia to bypass for /sdapi (delegate to the static api-auth) and healthcheck /,
and require LDAP auth (groups sd-read/sd-admin) for other resources.

Add docker network as kind on WSL2 doesn't preserve remote client IPs in current configuration.